refactor(agent-search): replace InputWithButton component with a simplified version in agent-search-input
