+++ web-app/src/app/(landing)/components/agent-search-input.tsx
@@ -1,7 +1,31 @@
 "use client";
 
-import InputWithButton from "@/components/input-with-button";
+import { ArrowUp } from "lucide-react";
+import { useRouter } from "next/navigation";
+import { useTranslations } from "next-intl";
+import type React from "react";
 
-export default function AgentSearchInput() {
-  return <InputWithButton />;
+import { Button } from "@/components/ui/button";
+
+export default function InputWithButton() {
+  const t = useTranslations("Landing.Page.Hero.AgentSearchInput");
+  const router = useRouter();
+  const handleClick = () => {
+    router.push(`/agents`);
+  };
+
+  return (
+    <>
+      <Button
+        onClick={handleClick}
+        size="lg"
+        className="flex items-center justify-between gap-2 rounded-lg"
+      >
+        <span className="text-muted-foreground">{t("placeholder")}</span>
+        <div className="bg-secondary rounded-md p-1">
+          <ArrowUp className="h-4 w-4" />
+        </div>
+      </Button>
+    </>
+  );
 }
